+++ src/modules/graphics/opengl.h
@@ -23,10 +23,3 @@
 
 #define GPU_SHADER_FIELDS \
   uint32_t program;
-
-#define GPU_TEXTURE_FIELDS \
-  bool native; \
-  uint8_t incoherent; \
-  GLuint id; \
-  GLuint msaaId; \
-  GLenum target;
