+++ src/modules/graphics/texture.h
@@ -1,6 +1,5 @@
 #include "data/textureData.h"
 #include "graphics/graphics.h"
-#include "graphics/opengl.h"
 #include "data/modelData.h"
 #include <stdbool.h>
 
@@ -15,27 +14,9 @@ typedef enum {
   TEXTURE_VOLUME
 } TextureType;
 
-typedef struct Texture {
-  TextureType type;
-  TextureFormat format;
-  uint32_t width;
-  uint32_t height;
-  uint32_t depth;
-  uint32_t mipmapCount;
-  CompareMode compareMode;
-  TextureFilter filter;
-  TextureWrap wrap;
-  uint32_t msaa;
-  bool srgb;
-  bool mipmaps;
-  bool allocated;
-  GPU_TEXTURE_FIELDS
-} Texture;
-
-Texture* lovrTextureInit(Texture* texture, TextureType type, struct TextureData** slices, uint32_t sliceCount, bool srgb, bool mipmaps, uint32_t msaa);
-Texture* lovrTextureInitFromHandle(Texture* texture, uint32_t handle, TextureType type, uint32_t depth);
-#define lovrTextureCreate(...) lovrTextureInit(lovrAlloc(Texture), __VA_ARGS__)
-#define lovrTextureCreateFromHandle(...) lovrTextureInitFromHandle(lovrAlloc(Texture), __VA_ARGS__)
+typedef struct Texture Texture;
+Texture* lovrTextureCreate(TextureType type, struct TextureData** slices, uint32_t sliceCount, bool srgb, bool mipmaps, uint32_t msaa);
+Texture* lovrTextureCreateFromHandle(uint32_t handle, TextureType type, uint32_t depth);
 void lovrTextureDestroy(void* ref);
 void lovrTextureAllocate(Texture* texture, uint32_t width, uint32_t height, uint32_t depth, TextureFormat format);
 void lovrTextureReplacePixels(Texture* texture, struct TextureData* data, uint32_t x, uint32_t y, uint32_t slice, uint32_t mipmap);
